create new pages, add images
Hello all,

I created a new page for the laptop category, though this will need remade in the future as right now it is more of a static page instead of a dynamically generated page.

I also made a product item page, which can be reused for all products I think. It is dynamically generated. I think that the category page needs to be made this way too.

I also added some javascript and css